refactor(aurix_mcu_src): hardware project template refactored and updated.
Updated the aurix_mcu_src project template to include also the build
Configuration for the Release mode. Also updated the hardware test
template to use the files in the project hardware template. By doing so
if in the future the build Configurations has to be updated you simply
need to update the Configurations in the project template without caring
about updating the tests