refactor(tests/car_component/pcu/pcu): Moved virtual Embeedded system into PCU.
In the car the Embeedded system is directly controlled by the pcu. So I
made the same thing in here by updating the logic code and virtual
emulation of pcu and Embeedded.

Finisged

Bugfixing

bugfix

fix tests compilation

fixed giei bug, pcu freq of as node to high => can2 full