fix(Build problem hardware): Added conditional compilation for 32-bit arch.
To fully compile the logic in the hardware the size of the object has to
be set accordingly. Until now the process has not been made and the
static check were not used in the hardware compilation process. Now the
debug checks has been added and also the conditional compilation to
differentiate the different sizes in 64 and 32 bit has been added.