fix(giei): removed GIEI_stop
Since in the check_running_condition function the control work already
set pos/neg torque to zero to set the control parameters. There is no
need to set them again, this may also cause conflict. Also since the
Also if the engines are set to 000 also the pos/neg torque are set to 0